Simplification of the rendering mechanism. User should use draw explicitly.
git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_3_0_0_dev@513 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/include/tulip/Curves.h
@@ -6,7 +6,7 @@
 
 namespace tlp {
   void curveVisibility(const Coord &startPoint,const std::vector<Coord> &bends, const Coord &endPoint,
-		       const Size &size, bool &polygon, bool &line, const GLdouble *transformMatrix, const GLint *viewportArray);
+		       const Size &size, bool &polygon, bool &line, const GLfloat *transformMatrix, const GLint *viewportArray);
   //====================================================================
   void polyLine(const std::vector<Coord> &,/* polyline vertces */
 		const Color &,             /* start color */
